pyproject.toml
[project]
name = "python-rag-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Vitalii Honchar",email = "weaxme@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain[openai] (>=0.3.25,<0.4.0)",
    "pypdf (>=5.4.0,<6.0.0)",
    "langchain-core (>=0.3.59,<0.4.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "langchain-postgres (>=0.0.14,<0.0.15)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "greenlet (>=3.2.2,<4.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "alembic (>=1.15.2,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
